Stretch-file-vanilla (#2694)

+++ tests/e2e/vsan_stretched_cluster_utils.go
@@ -53,6 +53,12 @@ import (
 	triggercsifullsyncv1alpha1 "sigs.k8s.io/vsphere-csi-driver/v3/pkg/internalapis/cnsoperator/triggercsifullsync/v1alpha1"
 )
 
+const (
+	filePath1  = "/mnt/volume1/file1.txt"
+	filePath2  = "/mnt/volume1/file2.txt"
+	accessMode = v1.ReadWriteMany
+)
+
 type FaultDomains struct {
 	primarySiteHosts   []string
 	secondarySiteHosts []string
@@ -220,10 +226,12 @@ func waitForAllNodes2BeReady(ctx context.Context, c clientset.Interface, timeout
 	if len(timeout) > 0 {
 		pollTime = timeout[0]
 	} else {
-		if os.Getenv("K8S_NODE_UP_WAIT_TIME") != "" {
+		if os.Getenv("K8S_NODES_UP_WAIT_TIME") != "" {
 			k8sNodeWaitTime, err := strconv.Atoi(os.Getenv(envK8sNodesUpWaitTime))
+			framework.Logf("k8sNodeWaitTime: %v, %T", k8sNodeWaitTime, k8sNodeWaitTime)
 			gomega.Expect(err).NotTo(gomega.HaveOccurred())
 			pollTime = time.Duration(k8sNodeWaitTime) * time.Minute
+			framework.Logf("pollTime: %v, %T", pollTime, pollTime)
 		} else {
 			pollTime = time.Duration(defaultK8sNodesUpWaitTime) * time.Minute
 		}
@@ -238,12 +246,19 @@ func waitForAllNodes2BeReady(ctx context.Context, c clientset.Interface, timeout
 			framework.Logf("error is %v", err)
 
 			if err != nil && !strings.Contains(err.Error(), "has prevented the request") &&
-				!strings.Contains(err.Error(), "TLS handshake timeout") {
+				!strings.Contains(err.Error(), "TLS handshake timeout") &&
+				!strings.Contains(err.Error(), "dial tcp") &&
+				!strings.Contains(err.Error(), ": EOF") {
 				return false, err
 			}
 			for _, node := range nodes.Items {
 				if !fnodes.IsConditionSetAsExpected(&node, v1.NodeReady, true) {
 					notReady = append(notReady, node)
+				}
+				for _, node := range nodes.Items {
+					if !fnodes.IsConditionSetAsExpected(&node, v1.NodeReady, true) {
+						notReady = append(notReady, node)
+					}
 				}
 			}
 			return len(notReady) == 0 && err == nil, nil
@@ -251,7 +266,6 @@ func waitForAllNodes2BeReady(ctx context.Context, c clientset.Interface, timeout
 	if len(notReady) > 0 {
 		return fmt.Errorf("not ready nodes: %v", notReady)
 	}
-
 	return err
 }
 
@@ -285,10 +299,18 @@ func deletePodsInParallel(ctx context.Context, client clientset.Interface, names
 func createPvcInParallel(ctx context.Context, client clientset.Interface, namespace string,
 	diskSize string, sc *storagev1.StorageClass,
 	ch chan *v1.PersistentVolumeClaim, lock *sync.Mutex, wg *sync.WaitGroup, volumeOpsScale int) {
+	var err error
+	var pvc *v1.PersistentVolumeClaim
 	defer ginkgo.GinkgoRecover()
 	defer wg.Done()
 	for i := 0; i < volumeOpsScale; i++ {
-		pvc, err := createPVC(ctx, client, namespace, nil, diskSize, sc, "")
+
+		//If file vanilla it uses ReadWriteMany
+		if rwxAccessMode {
+			pvc, err = createPVC(ctx, client, namespace, nil, diskSize, sc, accessMode)
+		} else {
+			pvc, err = createPVC(ctx, client, namespace, nil, diskSize, sc, "")
+		}
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 		lock.Lock()
 		ch <- pvc
@@ -400,12 +422,18 @@ func deletePvcInParallel(ctx context.Context, client clientset.Interface, pvclai
 // createPodsInParallel creates Pods in a given namespace in parallel
 func createPodsInParallel(client clientset.Interface, namespace string, pvclaims []*v1.PersistentVolumeClaim,
 	ctx context.Context, lock *sync.Mutex, ch chan *v1.Pod, wg *sync.WaitGroup, volumeOpsScale int) {
-
 	defer ginkgo.GinkgoRecover()
 	defer wg.Done()
+	var pod *v1.Pod
 
 	for i := 0; i < volumeOpsScale; i++ {
-		pod := fpod.MakePod(namespace, nil, []*v1.PersistentVolumeClaim{pvclaims[i]}, false, execCommand)
+		if rwxAccessMode {
+			pod = fpod.MakePod(namespace, nil, []*v1.PersistentVolumeClaim{pvclaims[i]}, false, "")
+
+		} else {
+			pod = fpod.MakePod(namespace, nil, []*v1.PersistentVolumeClaim{pvclaims[i]}, false, execCommand)
+
+		}
 		pod.Spec.Containers[0].Image = busyBoxImageOnGcr
 		pod, err := client.CoreV1().Pods(namespace).Create(ctx, pod, metav1.CreateOptions{})
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
@@ -413,6 +441,7 @@ func createPodsInParallel(client clientset.Interface, namespace string, pvclaims
 		ch <- pod
 		lock.Unlock()
 	}
+
 }
 
 // updatePvcLabelsInParallel updates the labels of pvc in a namespace in parallel
@@ -698,35 +727,77 @@ func createStsDeployment(ctx context.Context, client clientset.Interface, namesp
 // volumeLifecycleActions creates pvc and pod and waits for them to be in healthy state and then deletes them
 func volumeLifecycleActions(ctx context.Context, client clientset.Interface, namespace string,
 	sc *storagev1.StorageClass) {
-	pvc1, err := createPVC(ctx, client, namespace, nil, diskSize, sc, "")
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
 
-	pvs, err := fpv.WaitForPVClaimBoundPhase(ctx,
-		client, []*v1.PersistentVolumeClaim{pvc1}, framework.ClaimProvisionTimeout)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	volHandle := pvs[0].Spec.CSI.VolumeHandle
+	var pod1 *v1.Pod
 
-	pod1, err := createPod(ctx, client, namespace, nil, []*v1.PersistentVolumeClaim{pvc1}, false, execCommand)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	if vanillaCluster {
+		pvc1, err := createPVC(ctx, client, namespace, nil, diskSize, sc, "")
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 
-	vmUUID := getNodeUUID(ctx, client, pod1.Spec.NodeName)
-	framework.Logf("VMUUID : %s", vmUUID)
-	isDiskAttached, err := e2eVSphere.isVolumeAttachedToVM(client, volHandle, vmUUID)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	gomega.Expect(isDiskAttached).To(gomega.BeTrue(),
-		"Volume is not attached to the node volHandle: %s, vmUUID: %s", volHandle, vmUUID)
+		pvs, err := fpv.WaitForPVClaimBoundPhase(ctx,
+			client, []*v1.PersistentVolumeClaim{pvc1}, framework.ClaimProvisionTimeout)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+		volHandle := pvs[0].Spec.CSI.VolumeHandle
 
-	framework.Logf("Verify the volume is accessible")
-	_, err = e2eoutput.LookForStringInPodExec(namespace, pod1.Name,
-		[]string{"/bin/cat", "/mnt/volume1/fstype"}, "", time.Minute)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+		pod1, err = createPod(ctx, client, namespace, nil, []*v1.PersistentVolumeClaim{pvc1}, false, execCommand)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 
-	deletePodAndWaitForVolsToDetach(ctx, client, pod1)
+		vmUUID := getNodeUUID(ctx, client, pod1.Spec.NodeName)
+		framework.Logf("VMUUID : %s", vmUUID)
+		isDiskAttached, err := e2eVSphere.isVolumeAttachedToVM(client, volHandle, vmUUID)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+		gomega.Expect(isDiskAttached).To(gomega.BeTrue(),
+			"Volume is not attached to the node volHandle: %s, vmUUID: %s", volHandle, vmUUID)
 
-	err = fpv.DeletePersistentVolumeClaim(ctx, client, pvc1.Name, namespace)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	err = e2eVSphere.waitForCNSVolumeToBeDeleted(pvs[0].Spec.CSI.VolumeHandle)
-	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+		framework.Logf("Verify the volume is accessible")
+		_, err = e2eoutput.LookForStringInPodExec(namespace, pod1.Name,
+			[]string{"/bin/cat", "/mnt/volume1/fstype"}, "", time.Minute)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+		deletePodAndWaitForVolsToDetach(ctx, client, pod1)
+
+		err = fpv.DeletePersistentVolumeClaim(ctx, client, pvc1.Name, namespace)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+		err = e2eVSphere.waitForCNSVolumeToBeDeleted(pvs[0].Spec.CSI.VolumeHandle)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	}
+	if rwxAccessMode {
+		pvc1, err := createPVC(ctx, client, namespace, nil, "", sc, v1.ReadWriteMany)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+		time.Sleep(time.Duration(80) * time.Second)
+
+		pvs, err := fpv.WaitForPVClaimBoundPhase(ctx,
+			client, []*v1.PersistentVolumeClaim{pvc1}, framework.ClaimProvisionTimeout)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+		volHandle := pvs[0].Spec.CSI.VolumeHandle
+
+		verifyVolPropertiesFromCnsQueryResults(e2eVSphere, volHandle)
+
+		//Create Pod1
+		ginkgo.By(fmt.Sprintf("create pod with pvc: %s", pvc1.Name))
+		pod1, err := createPod(ctx, client, namespace, nil, []*v1.PersistentVolumeClaim{pvc1}, false, "")
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+		//Create file1.txt on Pod1
+		ginkgo.By("Create file1.txt on Pod1")
+		err = e2eoutput.CreateEmptyFileOnPod(namespace, pod1.Name, filePath1)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+		//Write data on file1.txt on Pod1
+		data := "This file file1 is written by Pod1"
+		ginkgo.By("Write on file1.txt from Pod1")
+		writeDataOnFileFromPod(namespace, pod1.Name, filePath1, data)
+
+		ginkgo.By(fmt.Sprintf("Deleting the pod : %s in namespace %s", pod1.Name, namespace))
+		err = fpod.DeletePodWithWait(ctx, client, pod1)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+		err = fpv.DeletePersistentVolumeClaim(ctx, client, pvc1.Name, namespace)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+		err = e2eVSphere.waitForCNSVolumeToBeDeleted(pvs[0].Spec.CSI.VolumeHandle)
+		gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	}
+
 }
 
 // scaleDownStsAndVerifyPodMetadata scales down replica of a statefulset if required
@@ -740,7 +811,6 @@ func scaleDownStsAndVerifyPodMetadata(ctx context.Context, client clientset.Inte
 		_, scaledownErr := fss.Scale(ctx, client, statefulset, replicas)
 		gomega.Expect(scaledownErr).NotTo(gomega.HaveOccurred())
 	}
-
 	fss.WaitForStatusReadyReplicas(ctx, client, statefulset, replicas)
 	ssPodsAfterScaleDown := fss.GetPodList(ctx, client, statefulset)
 	gomega.Expect(ssPodsAfterScaleDown.Items).NotTo(gomega.BeEmpty(),
